Refactored top-level kcm program into modules.
